+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/ChannelPostUpdate.kt
@@ -3,8 +3,9 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.update
 import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseSentMessageUpdate
 
 data class ChannelPostUpdate(
     override val updateId: UpdateIdentifier,
     override val data: Message
-) : BaseMessageUpdate
+) : BaseSentMessageUpdate