+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/EditChannelPostUpdate.kt
@@ -2,9 +2,9 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.update
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
-import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseEditMessageUpdate
 
 data class EditChannelPostUpdate(
     override val updateId: UpdateIdentifier,
     override val data: Message
-) : BaseMessageUpdate
+) : BaseEditMessageUpdate